requirements.txt
numpy
tqdm
torch
pytorch
pytorch-lightning
torch-scatter
torch-sparse
torch-geometric
ogb
torchmetrics
pyyaml
notebook
